allow $SEED to always set the initial state
While this contradicts the original behavior of the `MASTER` defined
symbol, this original behavior may be restored if `ORIGINAL_MASTER`
is also defined during compilation.

By default, `ORIGINAL_MASTER` is **NOT** defined and so setting
the `$SEED` environment variable at run time will set the
initial state of `rogue(6)` for debugging purposes.